requirements.txt
streamlit
fastapi
uvicorn
chromadb
openai
groq
tenacity
pandas
beautifulsoup4
requests
python-dotenv
sentence-transformers
python-multipart
